create and refactoring the listeners
